refactor: Merge utils.search_engine.filters and tils.search_engine.source_data back into utils.search_engine
